requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
paddlepaddle>=2.5.0
paddleocr>=2.7.0
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
python-docx>=1.1.0
pandas>=2.1.0
python-Levenshtein>=0.23.0
